+++ src/test/java/com/distasilucas/cryptobalancetracker/comparator/DescendingPercentageComparatorTest.java
@@ -4,6 +4,7 @@ import com.distasilucas.cryptobalancetracker.MockData;
 import com.distasilucas.cryptobalancetracker.model.response.crypto.CoinResponse;
 import org.junit.jupiter.api.Test;
 
+import java.math.BigDecimal;
 import java.util.Comparator;
 
 import static org.junit.jupiter.api.Assertions.assertAll;
@@ -17,9 +18,9 @@ class DescendingPercentageComparatorTest {
     void shouldCompareSuccessfully() {
         var coinInfo = MockData.getCoinInfo();
         var higherCoinResponse = MockData.getCoinResponse(coinInfo);
-        higherCoinResponse.setPercentage(60);
+        higherCoinResponse.setPercentage(BigDecimal.valueOf(60));
         var lowerCoinResponse = MockData.getCoinResponse(coinInfo);
-        lowerCoinResponse.setPercentage(40);
+        lowerCoinResponse.setPercentage(BigDecimal.valueOf(40));
 
         var firstCompare = comparator.compare(higherCoinResponse, lowerCoinResponse);
         var secondCompare = comparator.compare(lowerCoinResponse, higherCoinResponse);
